Move common rules logic to utils and constants modules
