+++ packages/eslint-plugin-kolibri/lib/rules/vue-no-unused-properties.js
@@ -8,29 +8,9 @@ const remove = require('lodash/remove');
 const eslintPluginVueUtils = require('eslint-plugin-vue/lib/utils');
 
 const utils = require('../utils');
+const constants = require('../constants');
 
-const GROUP_PROPERTY = 'props';
-const GROUP_DATA = 'data';
-const GROUP_COMPUTED_PROPERTY = 'computed';
-
-const PROPERTY_LABEL = {
-  [GROUP_PROPERTY]: 'property',
-  [GROUP_DATA]: 'data',
-  [GROUP_COMPUTED_PROPERTY]: 'computed property',
-};
-
-const reportUnusedProperties = (context, properties) => {
-  if (!properties || !properties.length) {
-    return;
-  }
-
-  properties.forEach(property => {
-    context.report({
-      node: property.node,
-      message: `Unused ${PROPERTY_LABEL[property.groupName]} found: "${property.name}"`,
-    });
-  });
-};
+const { GROUP_PROPS, GROUP_DATA, GROUP_COMPUTED } = constants;
 
 const create = context => {
   let hasTemplate;
@@ -56,7 +36,7 @@ const create = context => {
       unusedProperties = Array.from(
         eslintPluginVueUtils.iterateProperties(
           obj,
-          new Set([GROUP_PROPERTY, GROUP_DATA, GROUP_COMPUTED_PROPERTY])
+          new Set([GROUP_PROPS, GROUP_DATA, GROUP_COMPUTED])
         )
       );
 
@@ -70,7 +50,7 @@ const create = context => {
       });
 
       if (!hasTemplate && unusedProperties.length) {
-        reportUnusedProperties(context, unusedProperties);
+        utils.reportUnusedProperties(context, unusedProperties);
       }
     })
   );
@@ -88,7 +68,7 @@ const create = context => {
     },
     utils.executeOnRootTemplateEnd(() => {
       if (unusedProperties.length) {
-        reportUnusedProperties(context, unusedProperties);
+        utils.reportUnusedProperties(context, unusedProperties);
       }
     })
   );
